refactor: user 엔티티 경로이동
