+++ src/main/java/swyp_11/ssubom/domain/viewing/entity/PostView.java
@@ -2,7 +2,7 @@ package swyp_11.ssubom.domain.viewing.entity;
 
 import jakarta.persistence.*;
 import swyp_11.ssubom.domain.writing.entity.Post;
-import swyp_11.ssubom.global.security.entity.User;
+import swyp_11.ssubom.domain.user.entity.User;
 
 @Entity
 @Table(name = "PostView")
