+++ src/main/java/swyp_11/ssubom/domain/writing/entity/Reaction.java
@@ -1,7 +1,7 @@
 package swyp_11.ssubom.domain.writing.entity;
 
 import jakarta.persistence.*;
-import swyp_11.ssubom.global.security.entity.User;
+import swyp_11.ssubom.domain.user.entity.User;
 
 @Entity
 @Table(
