+++ src/main/java/swyp_11/ssubom/global/security/handler/CustomOAuthSuccessHandler.java
@@ -8,7 +8,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 import swyp_11.ssubom.global.security.dto.CustomOAuth2User;
-import swyp_11.ssubom.global.security.entity.User;
+import swyp_11.ssubom.domain.user.entity.User;
 import swyp_11.ssubom.global.security.jwt.JWTUtil;
 import swyp_11.ssubom.global.security.repository.UserRepository;
 import swyp_11.ssubom.global.security.service.RefreshTokenService;
