+++ src/main/java/swyp_11/ssubom/global/security/repository/UserRepository.java
@@ -2,7 +2,7 @@ package swyp_11.ssubom.global.security.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import swyp_11.ssubom.global.security.entity.User;
+import swyp_11.ssubom.domain.user.entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
